Add support for transforming SVG elements
Exact same syntax/semantics as other OpenSolid modules (rotateAround,
placeIn etc.), but maps to an SVG <g> element with a transform matrix.
This means that changing a transformation (modifying a rotation angle
etc.) will only affect the transformation matrix, not the underlying SVG
elements, so redrawing should be very fast.