Switch to getting render bounds from DOM
Necessary when the SVG diagram is *not* at the top-left of the page (does not start at (0,0)).

This requires the <svg> element to be tightly wrapped in a <div> since offsetLeft etc. are not available on SVGSVGElement objects, just HTMLElement ones. This could potentially be a breaking layout change, so should probably be rolled into a major version of opensolid/svg...or release opensolid/svg-interaction as a separate package with a separate render function.